# pkg/machine/e2e: Remove unnecessary copy of machine image.