Stop copying the pre-pulled uncompressed machine disk into the individual test dir. The machine pull code already makes a copy of the disk into the test's HOMEDIR/.local/share/containers/podman/machine, and works off that copy.

Before the change: `TESTDIR/<image>` is copied to `TESTDIR/podman_test/<image>` by the test, and then podman machine copies the image to `TESTDIR/podman_test/.local/share/containers/podman/machine/provider/<image>`

After the change: `TESTDIR/<image>` is copied to `TESTDIR/podman_test/.local/share/containers/podman/machine/provider/<image>` by podman machine

The image that is actually run is at `TESTDIR/podman_test/.local/share/containers/podman/machine/provider/<image>` in both instances.

<!--
Thanks for sending a pull request!

Please make sure you've read our contributing guidelines and how to submit a pull request (https://github.com/containers/podman/blob/main/CONTRIBUTING.md#submitting-pull-requests).

In case you're only changing docs, make sure to prefix the pull-request title with "[CI:DOCS]". That will prevent functional tests from running and save time and energy.

Finally, be sure to sign commits with your real name. Since by opening
a PR you already have commits, you can add signatures if needed with
something like `git commit -s --amend`.
-->

#### Does this PR introduce a user-facing change?

<!--
If no, just write `None` in the release-note block below. If yes, a release note
is required: Enter your extended release note in the block below. If the PR
requires additional action from users switching to the new release, include the
string "action required".

For more information on release notes, please follow the Kubernetes model:
https://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
None
```