@ashley-cui 2024-06-21
Initial local run of one test file, but lets see the performance in the full CI run.

Before change:
`make localmachine FOCUS_FILE=basic_test.go  53.08s user 41.46s system 32% cpu 4:51.64 total`

After change:
`make localmachine FOCUS_FILE=basic_test.go  36.01s user 19.68s system 28% cpu 3:17.05 total`