@Luap99 2024-06-21
Me wonders why `bud ...` tests are triggered on this PR. I would have assumed my new skip logic to skip bud here, i.e. on https://github.com/containers/podman/pull/23064 I see both machine and bud skipped so I assumed it works.

This PR touches machine so running machine test is correct but why would it trigger bud here?!