@Luap99 2024-06-21
Comparing to https://cirrus-ci.com/build/5300635074035712 it doesn't look like it made a difference for linux.
However macos was almost 60 mins there and 50 mins in your machine reset PR (https://github.com/containers/podman/pull/22941), here it is 34 mins so that looks like a very good speed up. I have not looked at macos timings much so I don't know the run to run variance but it looks like a good start.