@Luap99 2024-06-21
> > This PR touches machine so running machine test is correct but why would it trigger bud here?!
> 
> Base is very very old, does not include your changes

Well then I have a bug report for you :) You logformatter base commit is showing the wrong commit then, i.e. https://api.cirrus-ci.com/v1/artifact/task/5807831486562304/html/sys-remote-rawhide-root-host-sqlite.log.html lists 9ffac3317855 as base. Unless I am misunderstanding what "base" means. For me it would be last commit on the branch (PR) that is already in main.