@Luap99 2024-06-24
> HyperV: 10 min speedup Linux: 1-5 min speedup Mac: 20-30 min speedup WSL: 1-5 min speedup
> 
> Looks like it sped up the Mac tests a lot, which makes sense, since those machines were still slightly io bound. Ready for review!

At the least for the other PRs I checked I do not see that much speed up on linux/wsl. There is a lot of variance in the run two run timings so it is hard to tell of course. Of course this is a good to start but we have to get macos/hyperv to around 20 min to achieve our 30 min overall CI time goal.

Anyhow LGTM of course